KVM: x86/mmu: Get/Put TDP root page to be exported
Get/Put the root page of a KVM exported TDP page table based on KVM TDP
MMU.

When KVM TDP FD requests a TDP to export, it provides an address space
to indicate page roles of the TDP to export. In this RFC, KVM address space
0 is supported only. So, TDP MMU will select a root page role with smm=0
and guest_mode=0. (Level of root page role is from kvm->arch.maxphyaddr,
based on the assumption that vCPUs are homogeneous.)

TDP MMU then searches list tdp_mmu_roots for a existing root, or create a
new root if no one is found.
A specific kvm->arch.exported_tdp_header_cache is used to allocate the root
page in non-vCPU context.
The found/created root page will be marked as "exported".

When KVM TDP fd puts the exported FD, the mark of "exported" on root page
role will be removed.

No matter the root page role is exported or not, vCPUs just load TDP root
according to its vCPU modes.

In this way, KVM is able to share the TDP page tables in KVM address space
0 to IOMMU side.

                                         tdp_mmu_roots
                                             |
 role | smm | guest_mode              +------+-----------+----------+
------|-----------------              |      |           |          |
  0   |  0  |  0 ==> address space 0  |      v           v          v
  1   |  1  |  0                      |  .--------.  .--------. .--------.
  2   |  0  |  1                      |  |  root  |  |  root  | |  root  |
  3   |  1  |  1                      |  |(role 1)|  |(role 2)| |(role 3)|
                                      |  '--------'  '--------' '--------'
                                      |      ^
                                      |      |    create or get   .------.
                                      |      +--------------------| vCPU |
                                      |              fault        '------'
                                      |                            smm=1
                                      |                       guest_mode=0
                                      |
          (set root as exported)      v
.--------.    create or get   .---------------.  create or get   .------.
| TDP FD |------------------->| root (role 0) |<-----------------| vCPU |
'--------'        fault       '---------------'     fault        '------'
                                      .                            smm=0
                                      .                       guest_mode=0
                                      .
                 non-vCPU context <---|---> vCPU context
                                      .
                                      .

This patch actually needs to be split into several smaller ones.
It's tempted to be kept in a single big patch to show a bigger picture.
Will split them into smaller ones in next version.

Signed-off-by: Yan Zhao <yan.y.zhao@intel.com>