iommu/vt-d: Introduce require_clflush domain attribute
Introduce require_clflush domain attribute to indicate that CPU cache must
be flushed before adding/after removing a DMA page to/from the domain.

For domains with cache coherency not enforced, DMA is possible to be
non-coherent, meaning DMA can access the physical memory directly without
snooping CPU caches. So, when OS initializes a physical page and maps
the physical memory in IOMMU for DMA access, DMA is able to read stale
data in physical memory, bypassing OS's initialization data that may still
stay in CPU cacheline. This is deemed as information leak. [1].
So, clflush is required to make sure OS initialization data being flushed
into memory before DMA access is allowed.

The same applies to the DMA unmapping path where lacking of CPU cache flush
may allow a malicious guest to leak a memory/cache inconsistency situation
back to the host.

The require_clflush attribute is currently set to domains of type
IOMMU_DOMAIN_UNMANAGED  if cache coherency is not enforced in IOMMU.

For DMA domains leave the burden of flushing CPU cache to device driver
itself, as today.

Cc: Alex Williamson <alex.williamson@redhat.com>
Cc: Jason Gunthorpe <jgg@nvidia.com>
Cc: Kevin Tian <kevin.tian@intel.com>
Link: https://lore.kernel.org/lkml/20240109002220.GA439767@nvidia.com [1]
Suggested-by: Lu Baolu <baolu.lu@linux.intel.com>
Signed-off-by: Yan Zhao <yan.y.zhao@intel.com>