iommu/vt-d: Check if a PFN is for MMIO or not
Check if a PFN is for physical MMIO (i.e. not RAM) or not.

This check is mainly copied from kvm_is_mmio_pfn() in arch/x86/kvm/.

Unlike that in kvm_is_mmio_pfn(), ISA range is regarded as MMIO (PAT is WB
in host for ISA range). Though ISA range should not be mapped for VM usage,
still treat them as MMIO for code completeness.

Cc: Alex Williamson <alex.williamson@redhat.com>
Cc: Jason Gunthorpe <jgg@nvidia.com>
Cc: Kevin Tian <kevin.tian@intel.com>
Signed-off-by: Yan Zhao <yan.y.zhao@intel.com>